+++ src/main/java/com/account_catalogue/taxes/infraestructure/adapters/input/rest/data/response/TaxUpdateRes.java
@@ -12,7 +12,7 @@ public class TaxUpdateRes {
     private String idEnterprise;
     private String code;
     private String description;
-    private float interest;
+    private double interest;
     private Long refundAccount;
     private Long depositAccount;
     private Boolean status;
